Move ZmqMujinControllerClient into separate file
